Flask application file uploaded
This is the main Flask application file which contains the routes to different HTML pages. The various functions define the routes to various pages which should be landed upon clicking certain buttons or upon some action by the user . These actions are clearly defined in the various HTML pages uploaded in the templates folder.